and some transfer behaviour

+++ src/Basescape/TransferItemsState.cpp
@@ -610,7 +610,7 @@ void TransferItemsState::increaseByValue(int change)
 	{
 		double storesNeededPerItem = _game->getRuleset()->getItem(_items[getItemIndex(_sel)])->getSize();
 		double freeStores = _baseTo->getAvailableStores() - _baseTo->getUsedStores() - _iQty;
-		double freeStoresForItem = DBL_MAX;
+		double freeStoresForItem = (double)(INT_MAX);
 		if (!AreSame(storesNeededPerItem, 0.0))
 		{
 			freeStoresForItem = (freeStores + 0.05) / storesNeededPerItem;
